% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_hierarchical}
\alias{check_hierarchical}
\title{Helper functions.
check_hierarchical: Check if models given to r2mlm_comp are hierarchical.}
\usage{
check_hierarchical(modelA, modelB, calling_function, cluster_variable)
}
\arguments{
\item{modelA, modelB}{Models generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}.}

\item{calling_function}{Whether the helper funtion is r2mlm_comp_lme4 or
r2mlm_comp_nlme.}

\item{cluster_variable}{Clustering variable in dataframe.}
}
\description{
Helper functions.
check_hierarchical: Check if models given to r2mlm_comp are hierarchical.
}
