% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cwc}
\alias{get_cwc}
\title{get_cwc: Determine whether l1_vars are centered within cluster.}
\usage{
get_cwc(l1_vars, cluster_variable, data)
}
\arguments{
\item{l1_vars}{List of level 1 variables.}

\item{cluster_variable}{Clustering variable in dataframe.}

\item{data}{Dataset.}
}
\description{
get_cwc: Determine whether l1_vars are centered within cluster.
}
