% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_interaction_vars}
\alias{get_interaction_vars}
\title{get_interaction_vars: Return list of interaction variables}
\usage{
get_interaction_vars(model)
}
\arguments{
\item{model}{A model generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}, passed from the calling function.}
}
\description{
get_interaction_vars: Return list of interaction variables
}
