% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_random_slope_vars}
\alias{get_random_slope_vars}
\title{get_random_slope_vars: Return list of variable names with random effects}
\usage{
get_random_slope_vars(model, has_intercept, calling_function)
}
\arguments{
\item{model}{A model generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}, passed from the calling function.}

\item{has_intercept}{Whether or not the model has an intercept}

\item{calling_function}{Whether the helper funtion is r2mlm_lme4 or
r2mlm_nlme.}
}
\description{
get_random_slope_vars: Return list of variable names with random effects
}
