% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Pull and prepare data.}
\usage{
prepare_data(model, calling_function, cluster_variable, second_model = NULL)
}
\arguments{
\item{model}{A model generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}, passed from the calling function.}

\item{calling_function}{Whether the helper function is r2mlm_lme4 or
r2mlm_nlme.}

\item{cluster_variable}{Clustering variable in dataframe.}

\item{second_model}{A model generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}, passed from the calling function. Only used
when r2mlm_comp calls the helper function.}
}
\description{
First, pull dataframe associated with model using broom::augment. Then add
interaction terms and other user-created variables to the dataframe.
}
