% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2mlm_comp_manual.R
\name{r2mlm_comp_manual}
\alias{r2mlm_comp_manual}
\title{Compute R-squared values for two hierarchical multilevel models, with manual input.}
\usage{
r2mlm_comp_manual(
  data,
  within_covs_modA,
  between_covs_modA,
  random_covs_modA,
  gamma_w_modA,
  gamma_b_modA,
  Tau_modA,
  sigma2_modA,
  within_covs_modB,
  between_covs_modB,
  random_covs_modB,
  gamma_w_modB,
  gamma_b_modB,
  Tau_modB,
  sigma2_modB
)
}
\arguments{
\item{data}{Dataset with rows denoting observations and columns denoting
variables.}

\item{within_covs_modA, within_covs_modB}{List of numbers corresponding to the
columns in the dataset of the level-1 predictors used in the MLM (if none
used, set to NULL).}

\item{between_covs_modA, between_covs_modB}{List of numbers corresponding to
the columns in the dataset of the level-2 predictors used in the MLM (if
none used, set to NULL).}

\item{random_covs_modA, random_covs_modB}{List of numbers corresponding to the
columns in the dataset of the level-1 predictors that have random slopes in
the MLM (if no random slopes, set to NULL).}

\item{gamma_w_modA, gamma_w_modB}{Vector of fixed slope estimates for all
level-1 predictors, to be entered in the order of the predictors listed by
within_covs (if none, set to NULL).}

\item{gamma_b_modA, gamma_b_modB}{Vector of fixed intercept estimate (if
applicable; see has_intercept below) and fixed slope estimates for all
level-2 predictors, to be entered intercept first (if applicable) followed
by level-2 slopes in the order listed by between_covs (if none, set to
NULL).}

\item{Tau_modA, Tau_modB}{Random effect covariance matrix; note that the first
row/column denotes the intercept variance and covariances (if intercept is
fixed, set all to 0) and each subsequent row/column denotes a given random
slope’s variance and covariances (to be entered in the order listed by
random_covs).}

\item{sigma2_modA, sigma2_modB}{Level-1 residual variance.}
}
\value{
If inputs are valid models, then the output will be a list and
  associated graphical representation of R-squared decompositions.
}
\description{
\code{r2mlm_comp_manual} reads in raw data and multilevel model (MLM)
parameter estimates from two separate models under comparison (Model A and
Model B), and outputs all R-squared measures for both models. It also
produces side-by-side graphical comparisons of the R-squared measures for
Model A vs. Model B, that can be used to visualize changes in each measure
across models.
}
\details{
Details: assumes cwc, hierarchy/nesting
}
\examples{

modelA <- lmer(satisfaction ~ 1 + salary_c + control_c + salary_m +
control_m + s_t_ratio + (1 | schoolID), data = teachsat, REML = TRUE)

modelB <- lmer(satisfaction ~ 1 + control_c + control_m + s_t_ratio + (1
| schoolID), data = teachsat, REML = TRUE)

r2mlm_comp_manual(data = teachsat, within_covs_modA = c(5, 4),
between_covs_modA = c(7, 6, 8), random_covs_modA = NULL, gamma_w_modA =
c(1.52595, 2.66361), gamma_b_modA = c(19.68596, 1.43791, 3.65461, -0.36449),
Tau_modA = matrix(c(18.28349), 1, 1), sigma2_modA = 47.74, within_covs_modB =
c(4), between_covs_modB = c(6, 8), random_covs_modB = NULL, gamma_w_modB =
c(2.67877), gamma_b_modB = c(19.68596, 3.61712, -0.42340), Tau_modB =
matrix(c(26.77), 1, 1), sigma2_modB = 56.99)

}
\seealso{
Other r2mlm model comparison functions: 
\code{\link{r2mlm_comp}()}
}
\concept{r2mlm model comparison functions}
