% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covs}
\alias{get_covs}
\title{get_covs: Return the number in the dataframe associated with a given variable
name.}
\usage{
get_covs(variable_list, data)
}
\arguments{
\item{variable_list}{List of character names of variables.}

\item{data}{Dataset containing those variable names.}
}
\description{
get_covs: Return the number in the dataframe associated with a given variable
name.
}
