% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2mlm_comp.R
\name{r2mlm_comp}
\alias{r2mlm_comp}
\title{Compute R-squared values for two hierarchical multilevel models.}
\usage{
r2mlm_comp(modelA, modelB)
}
\arguments{
\item{modelA, modelB}{Models generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}.}
}
\value{
If inputs are valid models, then the output will be a list and
  associated graphical representation of R-squared decompositions. If models
  are not valid, it will return an error prompting the user to input valid
  models.
}
\description{
\code{r2mlm_comp} reads in two multilevel models (MLM) generated using
\code{\link[lme4]{lmer}} or \code{\link[nlme]{nlme}}, and outputs all
R-squared measures for both models. It also produces side-by-side graphical
comparisons of the R-squared measures for Model A vs. Model B, that can be
used to visualize changes in each measure across models.
}
\details{
#' Details: assumes cwc, hierarchy/nesting, both models lmer or both nlme, no mixing and matching
}
\seealso{
Other r2mlm model comparison functions: 
\code{\link{r2mlm_comp_manual}()}
}
\concept{r2mlm model comparison functions}
