% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2mlm_comp.R
\name{r2mlm_comp}
\alias{r2mlm_comp}
\title{Compute R-squared values for two hierarchical multilevel models.}
\usage{
r2mlm_comp(modelA, modelB)
}
\arguments{
\item{modelA, modelB}{Models generated using \code{\link[lme4]{lmer}} or
\code{\link[nlme]{nlme}}.}
}
\value{
If inputs are valid models, then the output will be a list and
  associated graphical representation of R-squared decompositions. If models
  are not valid, it will return an error prompting the user to input valid
  models.
}
\description{
\code{r2mlm_comp} reads in two multilevel models (MLM) generated using
\code{\link[lme4]{lmer}} or \code{\link[nlme]{nlme}}, and outputs all
R-squared measures for both models. It also produces side-by-side graphical
comparisons of the R-squared measures for Model A vs. Model B, that can be
used to visualize changes in each measure across models.
}
\details{
#' Details: assumes cwc, hierarchy/nesting, both models lmer or both nlme, no mixing and matching
}
\examples{
\dontrun{
# Using lme4 for your model

# Model A

modelA_lme4 <- lmer(satisfaction ~ 1 + salary_c + control_c + salary_m +
control_m + s_t_ratio + (1 | schoolID), data = teachsat, REML = TRUE)

# Model B, "salary" components removed

modelB_lme4 <- lmer(satisfaction ~ 1 + control_c + control_m + s_t_ratio + (1
| schoolID), data = teachsat, REML = TRUE)

r2mlm_comp(modelA_lme4, modelB_lme4)

# Using nlme for your model

# Model A

modelA_nlme <- lme(satisfaction ~ 1 + salary_c + control_c + salary_m +
                  control_m + s_t_ratio,
                  random = ~ 1 | schoolID,
                  data = teachsat,
                  method = "REML")

# Model B, "salary" components removed

modelB_nlme <- lme(satisfaction ~ 1 + control_c + control_m + s_t_ratio,
                  random = ~ 1 | schoolID,
                  data = teachsat,
                  method = "REML")

r2mlm_comp(modelA_nlme, modelB_nlme)
}


}
\seealso{
Other r2mlm model comparison functions: 
\code{\link{r2mlm_comp_manual}()}
}
\concept{r2mlm model comparison functions}
