% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2mlm_comp_manual.R
\name{r2mlm_comp_manual}
\alias{r2mlm_comp_manual}
\title{Compute R-squared values for two hierarchical multilevel models, with manual input.}
\usage{
r2mlm_comp_manual(
  data,
  within_covs_modA,
  between_covs_modA,
  random_covs_modA,
  gamma_w_modA,
  gamma_b_modA,
  Tau_modA,
  sigma2_modA,
  within_covs_modB,
  between_covs_modB,
  random_covs_modB,
  gamma_w_modB,
  gamma_b_modB,
  Tau_modB,
  sigma2_modB
)
}
\arguments{
\item{data}{Dataset with rows denoting observations and columns denoting
variables.}

\item{between_covs_modA, between_covs_modB}{List of numbers corresponding to the columns in the
dataset of the level-1 predictors used in the MLM (if none used, set to
NULL).}

\item{random_covs_modA, random_covs_modB}{List of numbers corresponding to the columns in the
dataset of the level-1 predictors that have random slopes in the MLM (if no
random slopes, set to NULL).}

\item{gamma_w_modA, gamma_w_modB}{Vector of fixed slope estimates for all level-1 predictors, to
be entered in the order of the predictors listed by within_covs (if none,
set to NULL).}

\item{gamma_b_modA, gamma_b_modB}{Vector of fixed intercept estimate (if applicable; see
has_intercept below) and fixed slope estimates for all level-2 predictors,
to be entered intercept first (if applicable) followed by level-2 slopes in
the order listed by between_covs (if none, set to NULL).}

\item{Tau_modA, Tau_modB}{Random effect covariance matrix; note that the first row/column
denotes the intercept variance and covariances (if intercept is fixed, set
all to 0) and each subsequent row/column denotes a given random slope’s
variance and covariances (to be entered in the order listed by
random_covs).}

\item{sigma2_modA, sigma2_modB}{Level-1 residual variance.}
}
\value{
If inputs are valid models, then the output will be a list and
  associated graphical representation of R-squared decompositions.
}
\description{
\code{r2mlm_comp_manual} reads in raw data and multilevel model (MLM)
parameter estimates from two separate models under comparison (Model A and
Model B), and outputs all R-squared measures for both models. It also
produces side-by-side graphical comparisons of the R-squared measures for
Model A vs. Model B, that can be used to visualize changes in each measure
across models.
}
\details{
Details: assumes cwc, hierarchy/nesting
}
\examples{

modelA_lme4 <- lmer(popular ~ 1 + extravCWC + texp + (extravCWC|class), data
= popularity, REML = TRUE)

modelB_lme4 <- lmer(popular ~ 1 + extravCWC + sexCWC + texp +
(extravCWC|class), data = popularity, REML = TRUE)

r2mlm_comp_manual(data = popularity, within_covs_modA = c(9),
between_covs_modA = c(5), random_covs_modA = c(9), gamma_w_modA = c(0.4985),
gamma_b_modA = c(4.4149, 0.0464), Tau_modA = matrix(c(0.5694, -0.0406,
-0.0406, 0.0326), 2, 2), sigma2_modA = 0.8921, within_covs_modB = c(9, 10),
between_covs_modB = c(5), random_covs_modB = c(9), gamma_w_modB = c(0.4562,
1.2396), gamma_b_modB = c(4.3757, 0.0491), Tau_modB = matrix(c(0.5833,
-0.0299, -0.0299, 0.0348), 2, 2), sigma2_modB = 0.5518)

}
\seealso{


Other r2mlm model comparison functions: 
\code{\link{r2mlm_comp}()}
}
\concept{r2mlm model comparison functions}
