% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2mlm_manual.R
\name{r2mlm_manual}
\alias{r2mlm_manual}
\title{Manually compute R-squared values for multilevel models.}
\usage{
r2mlm_manual(
  data,
  within_covs,
  between_covs,
  random_covs,
  gamma_w,
  gamma_b,
  Tau,
  sigma2,
  has_intercept = TRUE,
  clustermeancentered = TRUE
)
}
\arguments{
\item{data}{Dataset with rows denoting observations and columns denoting
variables.}

\item{between_covs}{List of numbers corresponding to the columns in the
dataset of the level-1 predictors used in the MLM (if none used, set to
NULL).}

\item{random_covs}{List of numbers corresponding to the columns in the
dataset of the level-1 predictors that have random slopes in the MLM (if no
random slopes, set to NULL).}

\item{gamma_w}{Vector of fixed slope estimates for all level-1 predictors, to
be entered in the order of the predictors listed by within_covs (if none,
set to NULL).}

\item{gamma_b}{Vector of fixed intercept estimate (if applicable; see
has_intercept below) and fixed slope estimates for all level-2 predictors,
to be entered intercept first (if applicable) followed by level-2 slopes in
the order listed by between_covs (if none, set to NULL).}

\item{Tau}{Random effect covariance matrix; note that the first row/column
denotes the intercept variance and covariances (if intercept is fixed, set
all to 0) and each subsequent row/column denotes a given random slope’s
variance and covariances (to be entered in the order listed by
random_covs).}

\item{sigma2}{Level-1 residual variance.}

\item{has_intercept}{If set to TRUE, the first element of gamma_b is assumed
to be the fixed intercept estimate; if set to FALSE, the first element of
gamma_b is assumed to be the first fixed level- 2 predictor slope; set to
TRUE by default.}

\item{clustermeancentered}{If set to TRUE, all level-1 predictors (indicated
by the within_covs list) are assumed to be cluster-mean-centered and
function will output all decompositions; if set to FALSE, function will
output only total decompositions (see Description above); set to TRUE by
default.}
}
\value{
If input is valid model, then the output will be a list and
  associated graphical representation of R-squared decompositions. If model
  is not valid, it will return an error.
}
\description{
\code{r2mlm_manual} takes as input raw data and parameter estimates from a
multilevel model, and outputs all relevant R-square measures and barchart
decompositions. Any number of level-1 and/or level-2 predictors is supported.
Any of the level-1 predictors can have random slopes.
}
\examples{
model <- lmer(popular ~ 1 + extravCWC + texp + (extravCWC|class), data =
popularity, REML = TRUE)

r2mlm_manual(popularity, within_covs = c(9), between_covs = c(5), random_covs
= c(9), gamma_w = c(0.4985), gamma_b = c(4.4149, 0.0464), Tau =
matrix(c(0.5694, -0.0406, -0.0406, 0.0326), 2, 2), sigma2 = 0.8921,
has_intercept = TRUE, clustermeancentered = TRUE))

}
\seealso{
\href{https://doi.org/10.1037/met0000184}{Rights, J. D., & Sterba,
  S. K. (2019). Quantifying explained variance in multilevel models: An
  integrative framework for defining R-squared measures. Psychological
  Methods, 24(3), 309–338.}

Other r2mlm single model functions: 
\code{\link{r2mlm}()}
}
\concept{r2mlm single model functions}
