% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teachsat}
\alias{teachsat}
\title{Teacher job satisfaction.}
\format{
A data frame with 2000 rows and 10 columns:
\describe{
  \item{schoolID}{school ID number}
  \item{teacherID}{teacher ID number}
  \item{satisfaction}{teacher job satisfaction}
  \item{control_c}{teacher control over curriculum, centered within school}
  \item{salary_c}{teacher salary, centered within school}
  \item{control_m}{teacher control over curriculum, school mean}
  \item{salary_m}{teacher salary, school mean, in thousands of dollars}
  \item{leadership}{school leadership competence}
  \item{type}{type of school: public, private, or charter}
}
}
\usage{
teachsat
}
\description{
A simulated dataset containing information about teacher job satisfaction.
Teachers clustered within schools.
}
\keyword{datasets}
